+++ test/dune/shims_test.exs
@@ -2,6 +2,7 @@ defmodule Dune.ShimsTest do
   use ExUnit.Case, async: true
 
   alias Dune.Success
+  alias Dune.Failure
 
   defmacrop sigil_E(call, _expr) do
     quote do
@@ -27,6 +28,26 @@ defmodule Dune.ShimsTest do
     test "decode atoms" do
       assert %Success{value: "json301", inspected: ~S("json301")} =
                ~E'JSON.decode!("\"json301\"")'
+    end
+  end
+
+  describe "iodata / chardata" do
+    test "List.to_string" do
+      assert %Success{value: <<1, 2, 3>>} = ~E'List.to_string([1, 2, 3])'
+      assert %Success{value: <<1, 2, 3>>} = ~E'List.to_string([1, [[2], 3]])'
+      assert %Success{value: "abc"} = ~E'List.to_string(["a", [["b"], "c"]])'
+
+      assert %Failure{message: "** (ArgumentError) cannot convert the given list" <> _} =
+               ~E'List.to_string([1, :foo])'
+    end
+
+    test "IO.chardata_to_string" do
+      assert %Success{value: <<1, 2, 3>>} = ~E'IO.chardata_to_string([1, 2, 3])'
+      assert %Success{value: <<1, 2, 3>>} = ~E'IO.chardata_to_string([1, [[2], 3]])'
+      assert %Success{value: "abc"} = ~E'IO.chardata_to_string(["a", [["b"], "c"]])'
+
+      assert %Failure{message: "** (ArgumentError) cannot convert the given list" <> _} =
+               ~E'IO.chardata_to_string([1, :foo])'
     end
   end
 end